refactor: inline page components into route files, remove performance mode navigation
- Inline DuplicateFinderPage and TagManagerPage into route files
- Inline SetlistsPage and SynthBackupsPage into route files with View named exports
- Remove onPlayInPerformanceMode prop and Performance Mode button from SetlistsPage
- Remove navigate/setActivePlaylistId calls from SetlistsPage route
- Update tests to import View components from route files
- Add routerFixture for route-level testing, AppSidebar browser tests
- Fix PresetsPage/SynthBackupsPage to remove unused setAppMode calls
