+++ src/features/synthBackups/components/SendEntryModal.tsx
@@ -1,7 +1,7 @@
 import SelectInput from "@/components/forms/SelectInput";
 import Button from "@/components/ui/Button";
 import Modal from "@/components/ui/Modal";
-import type { SendModalState } from "@/features/synthBackups/components/SynthBackupsPage.types";
+import type { SendModalState } from "@/routes/SynthBackupsPage";
 import type { SynthBackup } from "@/lib/collections/synthBackupManager";
 
 interface SendEntryModalProps {
